requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/in/base.in requirements/in/dev.in
#
--no-binary psycopg2

alabaster==0.7.12         # via sphinx
amqp==2.5.2               # via kombu
appdirs==1.4.3            # via black
argon2-cffi==19.2.0       # via -r requirements/in/base.in
asgiref==3.2.7            # via django
astroid==2.3.3            # via pylint, pylint-celery
attrs==19.3.0             # via black, pytest
babel==2.8.0              # via sphinx
backcall==0.1.0           # via ipython
bandit==1.6.2             # via -r requirements/in/dev.in
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements/in/dev.in
celery==4.4.2             # via -r requirements/in/base.in, django-celery-beat, django-celery-results, flower
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.0                # via -r requirements/in/base.in, black, django-click, pip-tools
coverage==5.1             # via django-coverage-plugin, pytest-cov
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via python3-openid
django-admin-tools==0.9.0  # via -r requirements/in/base.in
django-allauth==0.41.0    # via -r requirements/in/base.in
django-appconf==1.0.4     # via django-compressor
django-celery-beat==2.0.0  # via -r requirements/in/base.in
django-celery-results==1.2.1  # via -r requirements/in/base.in
django-click==2.1.0       # via -r requirements/in/base.in
django-compressor==2.4    # via -r requirements/in/base.in
django-coverage-plugin==1.8.0  # via -r requirements/in/dev.in
django-crispy-forms==1.9.0  # via -r requirements/in/base.in
django-debug-toolbar==2.2  # via -r requirements/in/dev.in
django-environ==0.4.5     # via -r requirements/in/base.in
django-extensions==2.2.9  # via -r requirements/in/dev.in
django-model-utils==4.0.0  # via -r requirements/in/base.in
django-redis==4.11.0      # via -r requirements/in/base.in
django-stubs==1.5.0       # via -r requirements/in/dev.in
django-timezone-field==4.0  # via django-celery-beat
django==3.0.5             # via -r requirements/in/base.in, django-allauth, django-appconf, django-celery-beat, django-debug-toolbar, django-model-utils, django-redis, django-stubs, django-timezone-field
docutils==0.16            # via sphinx
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via -r requirements/in/dev.in
faker==4.0.2              # via factory-boy
flake8-isort==2.9.1       # via -r requirements/in/dev.in
flake8==3.7.9             # via -r requirements/in/dev.in, flake8-isort
flower==0.9.4             # via -r requirements/in/base.in
gitdb==4.0.4              # via gitpython
gitpython==3.1.1          # via bandit
humanize==0.5.1           # via flower
idna==2.9                 # via requests
imagesize==1.2.0          # via sphinx
ipdb==0.13.2              # via -r requirements/in/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via ipdb
isort[pyproject]==4.3.21  # via flake8-isort, pylint
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via sphinx
kombu==4.6.8              # via celery
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, pylint
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements/in/dev.in, django-stubs
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.3           # via pytest, pytest-sugar, sphinx
parso==0.6.2              # via jedi
pathspec==0.8.0           # via black
pbr==5.4.5                # via stevedore
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.1.1             # via -r requirements/in/base.in
pip-tools==4.5.1          # via -r requirements/in/dev.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
psutil==5.6.7             # via purlovia
psycopg2==2.8.5           # via -r requirements/in/base.in
ptyprocess==0.6.0         # via pexpect
git+https://github.com/AngellusMortis/Purlovia.git@feature/config_root_logger#egg=purlovia  # via -r requirements/in/base.in
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.20           # via cffi
pydantic==1.4             # via purlovia
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython, sphinx
pylint-celery==0.3        # via -r requirements/in/dev.in
pylint-django==2.0.14     # via -r requirements/in/dev.in
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.4             # via pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytest-cov==2.8.1         # via -r requirements/in/dev.in
pytest-django==3.9.0      # via -r requirements/in/dev.in
pytest-dotenv==0.4.0      # via -r requirements/in/dev.in
pytest-sugar==0.9.2       # via -r requirements/in/dev.in
pytest==5.3.5             # via -r requirements/in/dev.in, pytest-cov, pytest-django, pytest-dotenv, pytest-sugar
python-crontab==2.4.1     # via django-celery-beat
python-dateutil==2.8.1    # via faker, python-crontab
python-dotenv==0.12.0     # via pytest-dotenv
python-slugify==4.0.0     # via -r requirements/in/base.in
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via babel, celery, django, django-timezone-field, flower
pyyaml==5.3.1             # via bandit, purlovia
rcssmin==1.0.6            # via django-compressor
redis==3.4.1              # via django-redis
regex==2020.4.4           # via black
requests-oauthlib==1.3.0  # via django-allauth
requests==2.23.0          # via django-allauth, purlovia, requests-oauthlib, sphinx
rjsmin==1.1.0             # via django-compressor
six==1.14.0               # via argon2-cffi, astroid, bandit, django-click, django-compressor, django-coverage-plugin, django-extensions, packaging, pip-tools, python-dateutil, stevedore, traitlets
smmap==3.0.2              # via gitdb
snowballstemmer==2.0.0    # via sphinx
sphinx==3.0.1             # via -r requirements/in/dev.in
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.3.1           # via django, django-debug-toolbar
stevedore==1.32.0         # via bandit
termcolor==1.1.0          # via pytest-sugar
testfixtures==6.14.0      # via flake8-isort
text-unidecode==1.3       # via faker, python-slugify
toml==0.10.0              # via black, isort
tornado==6.0.4            # via flower
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via django-stubs, mypy
urllib3==1.25.8           # via requests
vine==1.3.0               # via amqp, celery
wcwidth==0.1.9            # via prompt-toolkit, pytest
werkzeug==1.0.1           # via -r requirements/in/dev.in
whitenoise==5.0.1         # via -r requirements/in/base.in
wrapt==1.11.2             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
